refactor(endpointstate): extract per-endpoint ChainTracker into its own package
Prep for the ChainTracker & Probing Redesign (MAG-2171). Splits the
per-endpoint block-telemetry "data plane" out of rpcsmartrouter into a
dedicated package so the upcoming probing "decision plane" has a clean,
one-way boundary to build against (probing may read endpointstate; never
the reverse).

- Move endpoint_chain_tracker_manager.go -> endpointstate/endpoint_monitor.go
  (EndpointChainTrackerManager -> EndpointMonitor)
- Move endpoint_chain_fetcher.go -> endpointstate/endpoint_poller.go
  (EndpointChainFetcher -> EndpointPoller)
- Keep EndpointChainTrackerConfig / EndpointChainTrackerState names: they
  describe the underlying chaintracker concept, which is still accurate.
- Add protocol/probing/doc.go: empty decision-plane shell staking out the
  two-plane boundary for a later extraction ticket.
- Promote the in-package joinURLPath helper (surfaced as a cross-package
  dependency by the move) to common.JoinURLPath; relocate its test.
- Relocate the cancelFuncs lifecycle + ResetAllLatestBlocks count-correctness
  tests into endpointstate (they poke unexported maps, now in-package).
- Rewrite the rpcsmartrouter debug router-walk test to exported behavior
  only (real empty monitors + nil-safety); per-tracker count correctness is
  covered in endpointstate.

Verified: go build ./..., go vet ./..., endpointstate (+ -race), full
rpcsmartrouter suite, and common JoinURLPath all pass. One-way import edge
confirmed (rpcsmartrouter -> endpointstate); no cycle.
