+++ protocol/rpcsmartrouter/rpcsmartrouter_server_test.go
@@ -2322,104 +2322,6 @@ func TestSmartRouterSessionLeakPrevention_SingleProvider(t *testing.T) {
 }
 
 // ============================================================================
-// Tests for Epoch Cleanup Integration - EndpointChainTrackerManager Lifecycle
-// These tests validate that the EndpointChainTrackerManager properly cleans up
-// when trackers are removed (supporting epoch-based cleanup in RPCSmartRouter)
-// ============================================================================
-
-// TestEndpointChainTrackerManager_RemoveTrackerCallsCancel tests that RemoveTracker
-// properly invokes the cancel function for per-tracker context cancellation
-func TestEndpointChainTrackerManager_RemoveTrackerCallsCancel(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	t.Run("RemoveTracker invokes cancel function", func(t *testing.T) {
-		trackerManager := NewEndpointChainTrackerManager(ctx, EndpointChainTrackerConfig{
-			ChainID:          "ETH",
-			ApiInterface:     "jsonrpc",
-			AverageBlockTime: 12 * time.Second,
-			BlocksToSave:     10,
-		})
-		require.NotNil(t, trackerManager)
-		defer trackerManager.Stop()
-
-		// Manually add a cancel function to simulate a tracker
-		endpoint := "http://test:8545"
-		cancelCalled := false
-		trackerManager.cancelFuncs[endpoint] = func() { cancelCalled = true }
-
-		// Remove the tracker - should call cancel function
-		trackerManager.RemoveTracker(endpoint)
-
-		require.True(t, cancelCalled, "RemoveTracker should call the cancel function")
-		require.Empty(t, trackerManager.cancelFuncs)
-	})
-
-	t.Run("Stop invokes all cancel functions", func(t *testing.T) {
-		trackerManager := NewEndpointChainTrackerManager(ctx, EndpointChainTrackerConfig{
-			ChainID:          "ETH",
-			ApiInterface:     "jsonrpc",
-			AverageBlockTime: 12 * time.Second,
-			BlocksToSave:     10,
-		})
-		require.NotNil(t, trackerManager)
-
-		// Add multiple cancel functions
-		cancelledEndpoints := make(map[string]bool)
-		endpoints := []string{"http://ep1:8545", "http://ep2:8545", "http://ep3:8545"}
-
-		for _, ep := range endpoints {
-			trackerManager.cancelFuncs[ep] = func() { cancelledEndpoints[ep] = true }
-		}
-
-		// Stop should cancel all
-		trackerManager.Stop()
-
-		for _, ep := range endpoints {
-			require.True(t, cancelledEndpoints[ep], "Stop should cancel %s", ep)
-		}
-		require.Empty(t, trackerManager.cancelFuncs)
-	})
-
-	t.Run("concurrent RemoveTracker and Stop are thread-safe", func(t *testing.T) {
-		defer goleak.VerifyNone(t, goleak.IgnoreCurrent())
-
-		trackerManager := NewEndpointChainTrackerManager(ctx, EndpointChainTrackerConfig{
-			ChainID:          "ETH",
-			ApiInterface:     "jsonrpc",
-			AverageBlockTime: 12 * time.Second,
-			BlocksToSave:     10,
-		})
-		require.NotNil(t, trackerManager)
-
-		var wg sync.WaitGroup
-		const numGoroutines = 50
-
-		// Add many cancel functions
-		for i := 0; i < numGoroutines; i++ {
-			endpoint := fmt.Sprintf("http://endpoint%d:8545", i)
-			trackerManager.cancelFuncs[endpoint] = func() {}
-		}
-
-		// Simulate concurrent removal operations
-		for i := 0; i < numGoroutines; i++ {
-			wg.Add(1)
-			go func(id int) {
-				defer wg.Done()
-				endpoint := fmt.Sprintf("http://endpoint%d:8545", id)
-				trackerManager.RemoveTracker(endpoint)
-			}(i)
-		}
-
-		wg.Wait()
-
-		// Cleanup
-		trackerManager.Stop()
-		// If we reach here without race detector error or panic, the test passes
-	})
-}
-
-// ============================================================================
 // Mock Consistency for filterEndpointsByConsistency tests
 // ============================================================================
